[level 2] Title: 소수 찾기, Time: 13.77 ms, Memory: 84.2 MB -BaekjoonHub

+++ 프로그래머스/2/42839. 소수 찾기/소수 찾기.java
@@ -34,12 +34,13 @@ class Solution {
 
     public boolean isPrime(int val){
         if(val < 2) return false;
-        for(int i=2; i<= (int)Math.sqrt(val); i++){
+
+        for(int i=2; i<= Math.sqrt(val); i++){
             if(val % i == 0) {
                 return false;
             }
         }
-        return true; 
+        return true;
     }
 }
